backend/requirements.txt
fastapi
pymongo
python-dotenv
celery
redis
